backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-decouple>=3.8
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
django-storages>=1.14
boto3>=1.28
drf-spectacular>=0.26
python-decouple>=3.8
djangorestframework>=3.14
django-cors-headers>=4.0
pytz
djangorestframework-simplejwt
